Add IRCv3 tags to other events
We're already parsing these tags, but we were only putting them into
MessageEvent and PrivateMessageEvent. This just updates all the other
events to also store the tags.

I'm using this in combination with the `account-tag` capability to
associate user accounts to each of these events without needing to do a
whois. This is so much more efficient.